+++ node-js-peer/src/components/send-message.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from 'react'
 import { Text, Frame, Input } from 'react-curse'
 import { useLibp2pContext } from '../context/index.js'
 import { CHAT_TOPIC, PUBLIC_CHAT_ROOM_ID } from '../constants.js'
-import { layout, PositionProps } from '../index.js'
+import { layout, PositionProps } from '../lib/position.js'
 import { ChatMessage, useChatContext } from '../context/chat.js'
 import { peerIdFromString } from '@libp2p/peer-id'
 import { Logger } from '@libp2p/interface'
